Fix Family Member Medical History

+++ lib/screens/medication/add_medication_screen_simple.dart
@@ -110,7 +110,7 @@ class _AddMedicationScreenSimpleState extends State<AddMedicationScreenSimple> {
       isOngoing: _isOngoing,
       totalQuantity: _quantityController.text.isEmpty
           ? null
-          : int.parse(_quantityController.text),
+          : double.parse(_quantityController.text),
       instructions: _instructionsController.text.isEmpty
           ? null
           : _instructionsController.text.trim(),
